+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 ---------
 
 #### Next Release
+- Should dismiss dialog when translation presenter is stopped
 - Updated dependencies:
   - Grardle to 6.0